% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{association_file}
\alias{association_file}
\title{Example of association file.}
\format{
A data frame with 9 rows and 4 variables:
\describe{
  \item{ProjectID}{The ID of the project you are working on}
  \item{PoolID}{ID of the pool you are working on}
  \item{SubjectID}{ID of the subject}
  \item{CompleteAmplificationID}{ID of the current amplification}
}
}
\usage{
data("association_file")
}
\description{
Data obtained manually simulating real research data.
}
\keyword{datasets}
