% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ControlLine-class.R
\name{ControlLine}
\alias{ControlLine}
\title{R6 class representing a control cell line}
\description{
Single cell line characterized by a unique name and
a data frame of known integration sites
}
\details{
The known IS table must contain at least mandatory IS vars, as per options
set via `ISAnalytics::mandatory_IS_vars()`
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ControlLine-new}{\code{ControlLine$new()}}
\item \href{#method-ControlLine-clone}{\code{ControlLine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLine-new"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLine-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLine$new(name, known_iss)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{TODO}

\item{\code{known_iss}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
