% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_lines_management.R
\name{ControlLinesDb}
\alias{ControlLinesDb}
\title{R6 class representing a control cell line database}
\description{
The class acts as a database of all available control cell lines
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{default_lines}}{A list of ControlLine objects
provided as defaults. This property is read-only.}

\item{\code{user_defined_lines}}{A list of ControlLine objects
defined by users. This list is empty by default.
This property is read-only.}

\item{\code{available_lines}}{A list of ControlLine objects containing
all available control lines (default and user-defined).
This property is read-only.
Instantiates a new ControlLinesDb object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ControlLinesDb-new}{\code{ControlLinesDb$new()}}
\item \href{#method-ControlLinesDb-get_by_name}{\code{ControlLinesDb$get_by_name()}}
\item \href{#method-ControlLinesDb-add_cell_line}{\code{ControlLinesDb$add_cell_line()}}
\item \href{#method-ControlLinesDb-remove_cell_line}{\code{ControlLinesDb$remove_cell_line()}}
\item \href{#method-ControlLinesDb-edit_cell_line}{\code{ControlLinesDb$edit_cell_line()}}
\item \href{#method-ControlLinesDb-persist}{\code{ControlLinesDb$persist()}}
\item \href{#method-ControlLinesDb-clone}{\code{ControlLinesDb$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLinesDb-new"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLinesDb-new}{}}}
\subsection{Method \code{new()}}{
Upon initialization, the database tries to
load user-defined cell lines from the file system.
If the file does not exist, the database is initialized
with an empty list of user-defined cell lines.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLinesDb$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ControlLinesDb object
Retrieve a control cell line by name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLinesDb-get_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLinesDb-get_by_name}{}}}
\subsection{Method \code{get_by_name()}}{
This method returns a ControlLine object provided the
name exists in the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLinesDb$get_by_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the control cell line to retrieve}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ControlLine object
Add a new control cell line to the database
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLinesDb-add_cell_line"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLinesDb-add_cell_line}{}}}
\subsection{Method \code{add_cell_line()}}{
This method adds a new control cell line to the
database. The name of the new control cell line must be unique.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLinesDb$add_cell_line(
  name,
  known_iss,
  description = NULL,
  editable = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the new control cell line. Must be unique.}

\item{\code{known_iss}}{A data frame containing the known integration
sites of the new control cell line. The data frame must not be empty.}

\item{\code{description}}{A description of the new control cell line
(optional)}

\item{\code{editable}}{A boolean indicating whether the new control
cell line can be edited or not (optional, default is TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns TRUE if the add operation was successful,
FALSE otherwise.
Remove a control cell line from the database
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLinesDb-remove_cell_line"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLinesDb-remove_cell_line}{}}}
\subsection{Method \code{remove_cell_line()}}{
This method removes a control cell line from the
database. The name of the control cell line must exist in the
database (default cell lines can't be deleted).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLinesDb$remove_cell_line(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the control cell line to remove}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns TRUE if the remove operation was
successful,
FALSE otherwise.
Edit a control cell line
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLinesDb-edit_cell_line"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLinesDb-edit_cell_line}{}}}
\subsection{Method \code{edit_cell_line()}}{
This method edits a control cell line. The name of the
control cell line must exist in the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLinesDb$edit_cell_line(
  name,
  new_name = NULL,
  new_iss = NULL,
  new_desc = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the control cell line to edit}

\item{\code{new_name}}{The new name of the control cell line (optional)}

\item{\code{new_iss}}{A data frame containing the new known integration
sites of the control cell line (optional)}

\item{\code{new_desc}}{A new description of the control cell line (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns TRUE if the edit operation was successful,
FALSE otherwise.
Save the database to disk
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLinesDb-persist"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLinesDb-persist}{}}}
\subsection{Method \code{persist()}}{
This method saves the database to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLinesDb$persist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLinesDb-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLinesDb-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLinesDb$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
