% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{integration_matrix}
\alias{integration_matrix}
\title{Example of integration matrix.}
\format{
A data frame with 140 rows and 5 variables:
\describe{
  \item{chr}{The chromosome number}
  \item{integration_locus}{Number of the base at
 which the viral insertion occurred}
  \item{strand}{Strand of the integration}
  \item{CompleteAmplificationID}{ID of the current amplification}
  \item{Value}{Sequence count for the IS}
}
}
\usage{
data("integration_matrix")
}
\description{
Data obtained manually ismulating real research data.
}
\keyword{datasets}
