% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility-functions.R
\name{known_CEM_IS}
\alias{known_CEM_IS}
\title{Returns a table containing the known control integration sites.}
\usage{
known_CEM_IS()
}
\value{
Table of integration sites
}
\description{
This function builds a tribble which contains information
about the known integration sites belinging to control samples (CEMs).
It specifies the following fields: chromosome, integration locus,
strand, name (symbol) and strand of the closest gene to the IS.
}
\examples{
is <- known_CEM_IS()
is
}
