% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep-count-functions.R
\name{replicates_IS_count}
\alias{replicates_IS_count}
\title{Returns a dataframe containing the replicates count for
each integration and for each sample.}
\usage{
replicates_IS_count(
  af,
  matrix,
  subject_col = "SubjectID",
  field_sep = "_",
  amp_col = "CompleteAmplificationID",
  value_col = "Value"
)
}
\arguments{
\item{af}{The association file}

\item{matrix}{The integration matrix}

\item{subject_col}{The name of the subject column in af, 
default to "SubjectID"; if multiple columns identify the subject then
this must be a vector containing those columns names}

\item{field_sep}{The character that in control names separates the different 
columns values, default to "_"}

\item{amp_col}{The name of the amplificate column in af and matrix, 
default to "CompleteAmplificationID"}

\item{value_col}{The name of the SeqCount column in matrix, 
default to "Value"}
}
\value{
Dataframe of values corresponding to the replicate counts
}
\description{
This function counts the number of replicates in which each IS
is found and sums them to obtain an overall count for each sample.
A dataframe is returned: one row for each IS and one column for
each sample, the cell containing the overall replicate count
for that IS in that sample.
}
\examples{
data("association_file", package = "RISingVIS")
data("integration_matrix", package = "RISingVIS")
counts <- replicates_IS_count(association_file, integration_matrix)
head(counts)
}
