% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep-count-functions.R
\name{replicates_IS_ratio_byIS}
\alias{replicates_IS_ratio_byIS}
\title{Returns a dataframe containing the ratios for each IS between
the number of replicates for that IS in the control samples and
the number of replicates for that same IS in the other samples.}
\usage{
replicates_IS_ratio_byIS(
  af,
  matrix,
  subject_col = "SubjectID",
  amp_col = "CompleteAmplificationID",
  value_col = "Value",
  ctrl = "CEM37"
)
}
\arguments{
\item{af}{The association file}

\item{matrix}{The integration matrix}

\item{subject_col}{The name of the subject column in af,
default to "SubjectID"}

\item{amp_col}{The name of the amplificate column in af and matrix,
default to "CompleteAmplificationID"}

\item{value_col}{The name of the SeqCount column in matrix,
default to "Value"}

\item{ctrl}{The named list of control with known integration sites,
default is "CEM37" with known IS as in known_CEM_IS()}
}
\description{
This function counts the number of replicates in which each shared IS
is found and computes the ratio between controls and other samples.
This function focuses on two categories of shared IS: the ones that are
known to belong to controls and the one that come from the other samples.
Also it considers the count for single subjects and for the overall
sum of these subjects, meaning that the ratio is computed against
each subjects replicate count and against the sum of the replicate
count of all subjects.
After computation of replicate count gives as output a dataframe
containing all the ratios, which includes information about the IS,
the sample, the ratio itself and the type of IS considered.
}
\examples{
data("association_file", package = "RISingVIS")
data("integration_matrix", package = "RISingVIS")
R <- replicates_IS_ratio_byIS(association_file, integration_matrix)
head(R)
}
