% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-count-functions.R
\name{shared_CEM_known_IS_ratio}
\alias{shared_CEM_known_IS_ratio}
\title{Returns the ratio bewteen the SeqCount for shared known CEM IS
in the CEMs and the SeqCount for those shared IS
in the other samples}
\usage{
shared_CEM_known_IS_ratio(af, matrix)
}
\arguments{
\item{af}{The association file}

\item{matrix}{The integration matrix}
}
\value{
Single value of the ratio between SeqCounts
}
\description{
This function focuses on the shared IS that are known
to belong to CEMs.
After aggregation of the data gives as output
the ratio between the SeqCount for those IS in the CEMs and
the SeqCount for those IS in the other samples.
}
\examples{
data("association_file", package = "RISingVIS")
data("integration_matrix", package = "RISingVIS")
R <- shared_CEM_known_IS_ratio(association_file, integration_matrix)
R
}
