% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-count-functions.R
\name{shared_IS_ratio_byIS}
\alias{shared_IS_ratio_byIS}
\title{Returns a dataframe containing the ratios between the SeqCount
for each IS in the control samples and the SeqCount for the same IS
in the other samples.}
\usage{
shared_IS_ratio_byIS(
  af,
  matrix,
  subject_col = "SubjectID",
  field_sep = "_",
  amp_col = "CompleteAmplificationID",
  value_col = "Value",
  ctrl = "CEM37"
)
}
\arguments{
\item{af}{The association file}

\item{matrix}{The integration matrix}

\item{subject_col}{The name of the subject column in af,
default to "SubjectID"; if multiple columns identify the subject then
this must be a vector containing those columns names}

\item{field_sep}{The character that in control names separates the different
columns values, default to "_"}

\item{amp_col}{The name of the amplificate column in af and matrix,
default to "CompleteAmplificationID"}

\item{value_col}{The name of the SeqCount column in matrix,
default to "Value"}

\item{ctrl}{The named list of control with known integration sites,
default is "CEM37" with known IS as in known_CEM_IS()}
}
\value{
Dataframe of IS and values corresponding to the ratios of SeqCounts
}
\description{
This function returns the above mentioned ratio for each shared IS.
It focuses on two categories of shared IS: the ones that are
known to belong to controls and the one that come from the other samples.
Also it considers the count for single subjects and for the
overall sum of these subjects, meaning that the ratio is computed
against each subjects SeqCount and against the sum of the SeqCount
of all subjects.
After aggregation of the data gives as output a
dataframe containing all the ratios, which includes information about
the IS, the sample, the ratio itself and the type of IS considered.
}
\examples{
data("association_file", package = "RISingVIS")
data("integration_matrix", package = "RISingVIS")
R <- shared_IS_ratio_byIS(association_file, integration_matrix)
head(R)
}
