% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_lines_management.R
\name{ControlLine}
\alias{ControlLine}
\title{R6 class representing a control cell line}
\description{
Single cell line characterized by a unique name,
a data frame of known integration sites and optionally a description.
}
\examples{

## ------------------------------------------------
## Method `ControlLine$is_editable`
## ------------------------------------------------

cl <- ControlLine$new("name",
tibble::tibble(chr = c("1", "2"),
               integration_locus = c(102342, 4532231),
               strand = c("+", "-")))
cl
Renames the control line (if it is editable)

## ------------------------------------------------
## Method `ControlLine$rename`
## ------------------------------------------------

cl <- ControlLine$new("name",
tibble::tibble(chr = c("1", "2"),
               integration_locus = c(102342, 4532231),
               strand = c("+", "-")))
cl$rename(new_name = "new_name")
cl
Edits the known integration sites (if the control line is editable)

## ------------------------------------------------
## Method `ControlLine$edit_known_iss`
## ------------------------------------------------

cl <- ControlLine$new("name",
tibble::tibble(chr = c("1", "2"),
              integration_locus = c(102342, 4532231),
             strand = c("+", "-")))
cl$edit_known_iss(
tibble::tibble(chr = c("3", "4"),
               integration_locus = c(103536, 45353),
               strand = c("+", "-")))
cl
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{Name of the control line. This is a readonly field,
to rename the control line use the method `rename`.}

\item{\code{known_iss}}{Data frame containing the known integration sites.
This is a readonly field, to edit the known integration sites use
the method `edit_known_iss`.}

\item{\code{description}}{Description of the control line.
Instantiates a new ControlLine object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ControlLine-new}{\code{ControlLine$new()}}
\item \href{#method-ControlLine-print}{\code{ControlLine$print()}}
\item \href{#method-ControlLine-is_editable}{\code{ControlLine$is_editable()}}
\item \href{#method-ControlLine-rename}{\code{ControlLine$rename()}}
\item \href{#method-ControlLine-edit_known_iss}{\code{ControlLine$edit_known_iss()}}
\item \href{#method-ControlLine-clone}{\code{ControlLine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLine-new"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLine-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLine$new(name, known_iss, description = NULL, editable = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character string, name of the control line}

\item{\code{known_iss}}{A data frame containing the known integration sites.
Ensure this data frame does not contain duplicated rows.
Please note that `known_iss` cannot be empty.}

\item{\code{description}}{An optional character string, description of the
control line.}

\item{\code{editable}}{A logical, whether the control line is
editable or not.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ControlLine object
Prints the control line information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLine-print"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLine-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLine$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing
Returns whether the control line is editable or not
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLine-is_editable"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLine-is_editable}{}}}
\subsection{Method \code{is_editable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLine$is_editable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A logical
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cl <- ControlLine$new("name",
tibble::tibble(chr = c("1", "2"),
               integration_locus = c(102342, 4532231),
               strand = c("+", "-")))
cl
Renames the control line (if it is editable)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLine-rename"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLine-rename}{}}}
\subsection{Method \code{rename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLine$rename(new_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_name}}{A character string, new name of the control line}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cl <- ControlLine$new("name",
tibble::tibble(chr = c("1", "2"),
               integration_locus = c(102342, 4532231),
               strand = c("+", "-")))
cl$rename(new_name = "new_name")
cl
Edits the known integration sites (if the control line is editable)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLine-edit_known_iss"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLine-edit_known_iss}{}}}
\subsection{Method \code{edit_known_iss()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLine$edit_known_iss(new_known_iss)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_known_iss}}{A data frame containing the new known
integration sites. Ensure this data frame does not contain
duplicated rows.
Please note that `new_known_iss` cannot be empty.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{cl <- ControlLine$new("name",
tibble::tibble(chr = c("1", "2"),
              integration_locus = c(102342, 4532231),
             strand = c("+", "-")))
cl$edit_known_iss(
tibble::tibble(chr = c("3", "4"),
               integration_locus = c(103536, 45353),
               strand = c("+", "-")))
cl
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ControlLine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ControlLine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ControlLine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
