% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep-count-functions.R
\name{replicates_IS_count}
\alias{replicates_IS_count}
\title{Returns a dataframe containing the replicates count for
each integration and for each sample.}
\usage{
replicates_IS_count(af, matrix)
}
\arguments{
\item{af}{The association file}

\item{matrix}{The integration matrix}
}
\value{
Dataframe of values corresponding to the replicate counts
}
\description{
This function counts the number of replicates in which each IS
is found and sums them to obtain an overall count for each sample.
A dataframe is returned: one row for each IS and one column for
each sample, the cell containing the overall replicate count
for that IS in that sample.
}
\examples{
data("association_file", package = "RISingVIS")
data("integration_matrix", package = "RISingVIS")
counts <- replicates_IS_count(association_file, integration_matrix)
head(counts)
}
