% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-count-functions.R
\name{shared_control_known_IS_ratio}
\alias{shared_control_known_IS_ratio}
\title{Returns the ratio bewteen the SeqCount for shared known control IS
in all controls and the SeqCount for those shared IS
in the other samples}
\usage{
shared_control_known_IS_ratio(
  af,
  matrix,
  subject_col = "SubjectID",
  amp_col = "CompleteAmplificationID",
  value_col = "Value",
  ctrl = "CEM37"
)
}
\arguments{
\item{af}{The association file}

\item{matrix}{The integration matrix}

\item{subject_col}{The name of the subject column in af,
default to "SubjectID"}

\item{amp_col}{The name of the amplificate column in af and matrix,
default to "CompleteAmplificationID"}

\item{value_col}{The name of the SeqCount column in matrix,
default to "Value"}

\item{ctrl}{The named list of control with known integration sites,
default is "CEM37" with known IS as in known_CEM_IS()}
}
\value{
Single value of the ratio between SeqCounts
}
\description{
This function focuses on the shared IS that are known
to belong to controls.
After aggregation of the data gives as output
the ratio between the SeqCount for those IS in all controls and
the SeqCount for those IS in the other samples.
}
\examples{
data("association_file", package = "RISingVIS")
data("integration_matrix", package = "RISingVIS")
R <- shared_control_known_IS_ratio(association_file, integration_matrix)
R
}
